Minikube support and improved build tooling (#269)
* Add Minikube support.

* Force halt on certificate install failure.

* Unify build scripts and improve call usage.

Changelog:
	- Modified the <docker-build-e2e.sh> script so as to be able to invoke it from any directory whithin the project;
	- Change the shebang to directly execute the scripts;
	- Merged the Kind and Minikube fuctionality into the <cluster.sh> script;
	- Removed the <kind.sh> script;
	- Modified the <clean_reinstall.sh> script to reflect the new usage;
	- Added usage information in the scripts <docker-build-e2e.sh> and <cluster.sh>;
	- Added a descriptive help message in the <cluster.sh> script;

The new <cluster.sh> script covers not only the integrated functionality for the creation and preparation of both Kind and Minikube clusters, but also allows the invocation of the <docker-build-e2e.sh> script through the <b> flag. Thus decreasing the amount of steps to setup and run the project.

* Add new make targets and improve the <cluster.sh> script.

Changelog:
	- Fix tag and host parameter order in the <cluster.sh> script;
	- Building a cluster can now be done through make calls;
	- Make can also install Undistro into the created cluster;
	- Update Github workflow file to call the new script;

The new make target <cli-kind_install> and <cli-minikube_install> perform all the required operations to have Undistro running. These compile the CLI, setup a cluster and a registry, and lastly install Undistro in it.

Two other targets were added to perform the first steps without installing Undistro. Thus, the targets <setup_kind> and <setup_minikube> only take care of preparing both cluster and registry.

The behavior of such targets can be customized with some environment variables as follows:
	REG_HOST: The host address of the registry, defaulting to "localhost" if not set;
	UND_IMG_TAG: The Docker tag used in the Undistro image, defaulting to "latest";
	UND_CONF: The name of Undistro's configuration file, defaulting to "undistro.yaml";
	MINIK_IP: The IP address of the Minikube runtime;

Disclaimer: When creating a Minikube cluster, the registry host must have the same address as the Minikube's runtime.

Usage examples:

To setup a KinD cluster with a local registry, compile Undistro and install it with the tag "test" and the configuration file "conf.yml":
	$ make cli-kind_install UND_IMG_TAG=test UND_CONF=conf.yml

To only setup a minikube cluster and registry:
	$ make setup_minikube $MINIK_IP=192.168.39.2 $REG_HOST=192.168.39.2