(WIP) Began implementation of Azure specific functions and integration with the provider's SDK.
Changes:
	- Added Azure to the <providerInfo> variable in the <install.go> file;
	- Imported Azure SDK;
	- Implemented many provider specific functions in the <azure> package;
	- Fix invalid cluster template CIDR reference;
	- Initiated the implementation of the metadata funcinality for Azure;
	- Moved the <GetData> and <RemoveDuplicateNetwork> functions into the <util> package;
	- Created a pro

Azure requires users of its SDK to authenticate with it. Environment variables containing credentials informed through Undistro's configuration file are exported in the <azure> package, later being consumed by the <NewAuthorizerFromEnvironment> function from the SDK.

The <GetData> and <RemoveDuplicateNetwork> functions were moved from the <aws> package due to its generic nature, since they apply to multiple providers, including Azure.

What's left:
	- Testing of the <undistro-azure> chart;
	- Testing of the <getLocations> function since it uses the Azure SDK;
	- Creation of the metadata description YAML files;
	- Review of <azuremachine_types.go>;
	- Finish the <GetFlavors> function implementation;
	- Review the metadata functions in the <azure> package;

The <GetFlavors> function intends to dynamically obtain a list of Kubernetes versions supported by AKS, by calling Microsoft's REST API. This call requires at least the API version and the subscription ID to be passed with the request, but the location and the resource type can be passed as well. It may be possible to use this same API to get a list of versions for the VM flavor as well, possibly by stipulating the resource type parameter.

Some documentation of the API can be found in the following links:
	https://docs.microsoft.com/en-us/rest/api/aks/managed-clusters/get
	https://docs.microsoft.com/en-us/rest/api/aks/managed-clusters/list

Another way to obtain such list is by using Azure's CLI, which requires a location to be informed:
	$ az aks get-versions --location <LOCATION_NAME>

More details are available in the link bellow:
	https://docs.microsoft.com/en-us/cli/azure/aks?view=azure-cli-latest#az_aks_get_versions