[skip e2e] Multi local cluster refac (#324)
* Added <clean> make target.

This new target removes cached Helm tarballs from the <pkg/fs/charts> directory. The target has been incorporated into the install targets as well as the <run> and <build> ones as a dependency.

* Refactored the handling of local cluster detection.

Changes:
	- Modified the <util.isLocalCluster> function to primarily return
	an integer;
	- Removed <echo> calls from the <cluster.sh> script;
	- Renamed the <validateKindConfig> function to
	<validateLocalClusConfig> from <cli> package;
	- Added a <switch> in the <cli.validateLocalClusConfig> function;

The new return type from the <isLocalCluster> function possibilitates
its usage not only to detect whether a cluster is local or remote, but
also which type of local cluster is being used. This is done by mapping
each type to a number as follows:

	RETURNED VALUE		LOCAL CLUSTER TYPE
	-1 					Not a local cluster
	0					Error in obtaining the list of containers
	1 					Local Kind cluster
	2 					Local Minikube cluster

This approach eases the process of adding new types of local cluster,
wherefrom a new integer would suffice to identify it.

Currently it's important to identify which kind of local cluster engine
is being used, so to query the correct IP address for MetalLB's
configuration.

* Removed debug print statement of container images and added more documentation comments.