Cluster deletion never ending (#417)
There was a conflict in some of the UnDistroCluster states, in our admission webhooks there was a validation that didn't allow any update in the cluster state if it wasn't with `Ready = true`. When the cluster went into the delete state, the ready went to false and then the controller looped and never got out of the reconciliation loop. Added an exception in case the cluster is being deleted in our admission webhooks checks and some aesthetic refactors.